move entity sentiment to v1
