+++ language/cloud-client/v1beta2/snippets_test.py
@@ -69,26 +69,6 @@ def test_syntax_file(capsys):
     assert 'NOUN: President' in out
 
 
-def test_sentiment_entities_text(capsys):
-    snippets.entity_sentiment_text(
-        'President Obama is speaking at the White House.')
-    out, _ = capsys.readouterr()
-    assert 'Content : White House' in out
-
-
-def test_sentiment_entities_file(capsys):
-    snippets.entity_sentiment_file(TEST_FILE_URL)
-    out, _ = capsys.readouterr()
-    assert 'Content : White House' in out
-
-
-def test_sentiment_entities_utf(capsys):
-    snippets.entity_sentiment_text(
-        'foo→bar')
-    out, _ = capsys.readouterr()
-    assert 'Begin Offset : 4' in out
-
-
 def test_classify_text(capsys):
     snippets.classify_text(
         'Android is a mobile operating system developed by Google, '
